Update Game_Enemy_Extension.rb
Made it so the stats will scale with their level. The scaling factor can be change. 10 feels right for main stats, but for minor stat growth would need to be a higher value. (basically, it the growth step wherever 10 (or the set value) levels is reach, the stat grows another full version of its base self. currently it is offset by one level to allow level 1 to use the base stats)