Update Calculations_Formulas.rb
Added a magic attack that add the power to the scaled magic variable. This allows a fix amount without depending on the magic value (like magic attack) early on but will grow in power as long as magic is greater than 0. (Note: all scale factors are placeholders unless I test it enough)