Update Event_Triggers.rb
Added selfswitch, and init listening. Init is designed to be called once the event is created and thus events would need to add `*init` to their name to make use of it. selfswitch is an extra channel to listen too for self switch changes, but it will be called when any self switch change happens. Note: most of this auto functionality happens in the _Inj file and can be disable or modified there.

Also added the ability to make triggers for manual calling. Currently limited to around 7, but MAX_TRIGGERS can be increased. Since it uses an array to keep track of events, limiting it can help accidently adding hundreds of nil elements (like naming an event with `*trig85345`. Also can register events by adding `*trig#` (where # is the trigger id) in the event name.
Note: this system was built to work not on autorun or parallel events since they do not need remote calls.