Create Battle_Positions.rb
Added a concept of battle positions. It currently only reduce damage base on distance. It also needs functions that do not exist in the base classes. These functions are to allow aliasing without breaking the logic, but will need to be injected. I will update the ext classes with these functions