Update Note_Reader.rb
Added tagging and included a priority variable for baseItem. Tags are not used at the moment, can be used for filtering or advance sorting. Priority allow ordering items displayed in lists. This is to allow not needing to reorganize the database. The current way it is applied is a bit limiting but preserves item id in the same priority without having to presort the array by id. Priority can be overridden or alias to add custom rating calculations. \@priority is used as a catch so calculations based on static values won't need to run more than once a session.