Update Note_Reader.rb
Added the ability to check items for tags in the interpreter...or well the ability to see if enough items have that tag. It is mostly for trigger events where more than one items could be used and should not be ran every update since it could loop the whole inventory for false cases.