Create Event_Triggers.rb
Move Event_Triggers here and change it to handle a hash for event ids. This allow variable 1 to be listen to instead of responding to all variables state changes. adding *var1 to the event name should assign to var 1 state changes, but *var can be used to listen to all changes, but this is unadvised since blacklisting is now removed (since it could cause events listening to *var to start every update if a var changes every update).