Update Calculations_Formulas.rb
change some of the math. Set the default max modifier to max stat of 1000 instead of 255. power_attack still need to be look into, but for now it mirrors magic attack, but with a focus on level instead of power (base power is between 1-5). I may need to redo and simplify these since they follow a value modify by a percent and some kind of amplifier.
I kind of did not like how easy it was for def to remove damage from an attack, so that the reason for switching to a scaled approach base on attack and def.