Update Field_Actions.rb
Ran into an issue where jump only work if there enough space ahead of the landing point. I assume it was direction and the reason why I pass direction was because 0 allow jumping anytime. I decided to try 1 since I assume 0 cancel the points. So far it works, but I am not sure if it is correct. I hope it is treated as here and not some odd direction.