Update Field_Actions.rb
Added a preparing jump state where when set with an array of [follower visibility, jump x, jump y, stage] it will either jump is follower visibility == false or will run through stages to gather, hide followers, jump and force followers to move to jump point, and make followers visible. This is mostly to prevent followers from walking across impassible tiles, but can be a bit slow waiting for them so I may add a way to not gather and just relocate or make them move faster.