Update LocalizationEvents.cs
I figure out more about handling input and managing the message. It lacks a way to stop the typing effect and display the full text and it still do not have a language. 

Also, the default types like file won't work unless the path is strip since both icon and picture append the pass value to the directory they look in. Instead, I am using a string and expect the user to type the file name with extension(.png)

Yeah I have to edit the desc and name. This is all experiments.