Update CustomEvents.cs
Added a way to check map events. This check would be ideal for handling field actions and may be move to a such a dedicated addon on if I decided to add more features. The only issue is that it needs to set something (using a switch) instead of being a conditional.