Update CustomEvents.cs
Added a jump action that will jump in the character current direction as long as they can move to the tile. I may modify it later to allow not allow collision checks, enable or disable vehicle, and maybe allow a vehicle id to fake move without being in the vehicle.