Update CustomEvents.cs
Updated the jump to have more options like to allow nav on areas with region id or tag (I should handle tag as a bitflag in the future and rename it tagID or something). I was going to use vehicles as a nav option, but was unable to figure out how to get a uid needed without changing the source code. I may add jump distance next, but it may not have a collision check (aka jump over). I also may expose the collision check to events for more control.