Update CustomEvents.cs
Added a way get the x and y form a point base on the direction of an event. Later I play to add a selection to check the trigger action since I am using it to disable the field action when there is an action event Infront of the player.